Add properties for svn hotcopy options.
https://marc.info/?l=amanda-hackers&m=150430373325284

+++ application-src/amsvnmakehotcopy.pl
@@ -57,15 +57,22 @@ sub new {
 	    'script requires svnrepository property');
     }
 
+    $self->{'incremental'} = $self->{'options'}->{'incremental'};
+
     return $self;
 }
 
 sub declare_options {
     my ( $class, $refopthash, $refoptspecs ) = @_;
     $class->SUPER::declare_options($refopthash, $refoptspecs);
-    push @$refoptspecs, ( 'svnadminexecutable=s', 'svnrepository=s' );
+    push @$refoptspecs, ( 'svnadminexecutable=s', 'svnrepository=s',
+			  'clean-logs=s', 'incremental=s' );
 
     $class->store_option($refopthash, 'svnadminexecutable', 'svnadmin');
+    $class->store_option($refopthash,
+	'clean-logs', $class->boolean_property_setter($refopthash));
+    $class->store_option($refopthash,
+	'incremental', $class->boolean_property_setter($refopthash));
 }
 
 sub command_pre_dle_estimate {
@@ -73,12 +80,16 @@ sub command_pre_dle_estimate {
 
     my $repo = $self->{'svnrepository'};
     my $dst = $self->{'options'}->{'device'};
+    my @opts;
+
+    push @opts, '--clean-logs' if $self->{'options'}->{'clean-logs'};
+    push @opts, '--incremental' if $self->{'incremental'};
 
     make_path($dst);
-    remove_tree($dst, {keep_root => 1});
+    remove_tree($dst, {keep_root => 1}) unless $self->{'incremental'};
 
     my $rslt = system {$self->{'svnadminexecutable'}} (
-        'svnadmin', 'hotcopy', $repo, $dst );
+	'svnadmin', 'hotcopy', @opts, '--', $repo, $dst );
 }
 
 package main;
